Fix cromwell_root/script execution issues
This script wasn't being executed because it was disabled to
avoid the volume mounting issue (fixed by #56).

There was another uncovered problem, the script content was
trying to deal with gcp files instead of the local file system,
which is solved now.

Let's enable the script execution so that delocalization works.

NOTE: The delocalization steps involving /google are disabled.